Fix tests after Flutter 3.10 upgrade
