+++ lib/widgets/charts/dashboard_measurables_line_chart.dart
@@ -1,4 +1,5 @@
 import 'dart:core';
+import 'dart:math';
 
 import 'package:enum_to_string/enum_to_string.dart';
 import 'package:flutter/material.dart';
@@ -146,7 +147,7 @@ class MeasurablesChartInfoWidget extends StatelessWidget {
           top: 0,
           left: 10,
           child: SizedBox(
-            width: MediaQuery.of(context).size.width - 20,
+            width: max(MediaQuery.of(context).size.width, 300) - 20,
             child: Row(
               children: [
                 ConstrainedBox(
